pyproject.toml
[tool.poetry]
name = "tcc-mayumi"
version = "0.1.0"
description = ""
authors = ["Alison Almeida <wsalisonxp@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sanic = "^23.6.0"
gmqtt = "^0.6.12"
uvloop = "^0.19.0"
python-socketio = "^5.10.0"
motor = "^3.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
